[NFC] Move util/InformativeMailNotifier to reporters/utils
Note we plan to add LLVMCommentFormatter, LLVMFailBuildGenerator and LLVMFailGitHubReporter soon.
reporters/* seems the right place for them.
