+++ buildbot/osuosl/master/config/status.py
@@ -4,7 +4,7 @@ from buildbot.reporters.generators.build import BuildStartEndStatusGenerator
 from twisted.python import log
 
 import config
-from zorg.buildbot.util.InformativeMailNotifier import (
+from zorg.buildbot.reporters.utils import (
     LLVMInformativeMailGenerator,
     LLVMDefaultBuildStatusGenerator
 )
